feat: move PublicUser para o pacote de DTO
